[FLINK-22657][Connectors/Hive] HiveParserDDLSemanticAnalyzer directly return operations
