--- flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/desc/CreateTableASDesc.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.table.planner.delegation.hive.desc;
-
-import org.apache.flink.table.planner.delegation.hive.copy.HiveParserASTNode;
-
-import java.io.Serializable;
-
-/** A desc for CTAS. */
-public class CreateTableASDesc implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private final HiveParserCreateTableDesc createTableDesc;
-    private final HiveParserASTNode query;
-
-    public CreateTableASDesc(HiveParserCreateTableDesc createTableDesc, HiveParserASTNode query) {
-        this.createTableDesc = createTableDesc;
-        this.query = query;
-    }
-
-    public HiveParserCreateTableDesc getCreateTableDesc() {
-        return createTableDesc;
-    }
-
-    public HiveParserASTNode getQuery() {
-        return query;
-    }
-}
