--- flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/desc/HiveParserDropDatabaseDesc.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.table.planner.delegation.hive.desc;
-
-import java.io.Serializable;
-
-/** Desc for drop database. */
-public class HiveParserDropDatabaseDesc implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    private final String databaseName;
-    private final boolean ifExists;
-    private final boolean cascade;
-
-    public HiveParserDropDatabaseDesc(String databaseName, boolean ifExists, boolean cascade) {
-        this.databaseName = databaseName;
-        this.ifExists = ifExists;
-        this.cascade = cascade;
-    }
-
-    public String getDatabaseName() {
-        return databaseName;
-    }
-
-    public boolean ifExists() {
-        return ifExists;
-    }
-
-    public boolean cascade() {
-        return cascade;
-    }
-}
