+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/parse/HiveParserDDLSemanticAnalyzer.java
@@ -18,34 +18,87 @@
 
 package org.apache.flink.table.planner.delegation.hive.parse;
 
+import org.apache.flink.sql.parser.hive.ddl.HiveDDLUtils;
+import org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabase;
+import org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable;
+import org.apache.flink.table.api.TableColumn;
+import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.ValidationException;
+import org.apache.flink.table.api.WatermarkSpec;
+import org.apache.flink.table.api.constraints.UniqueConstraint;
+import org.apache.flink.table.catalog.Catalog;
+import org.apache.flink.table.catalog.CatalogBaseTable;
+import org.apache.flink.table.catalog.CatalogDatabase;
+import org.apache.flink.table.catalog.CatalogDatabaseImpl;
+import org.apache.flink.table.catalog.CatalogFunction;
+import org.apache.flink.table.catalog.CatalogFunctionImpl;
+import org.apache.flink.table.catalog.CatalogManager;
+import org.apache.flink.table.catalog.CatalogPartition;
+import org.apache.flink.table.catalog.CatalogPartitionImpl;
 import org.apache.flink.table.catalog.CatalogPartitionSpec;
+import org.apache.flink.table.catalog.CatalogTable;
+import org.apache.flink.table.catalog.CatalogTableImpl;
+import org.apache.flink.table.catalog.CatalogView;
+import org.apache.flink.table.catalog.CatalogViewImpl;
+import org.apache.flink.table.catalog.FunctionLanguage;
+import org.apache.flink.table.catalog.ObjectIdentifier;
 import org.apache.flink.table.catalog.ObjectPath;
+import org.apache.flink.table.catalog.UnresolvedIdentifier;
+import org.apache.flink.table.catalog.exceptions.DatabaseNotExistException;
 import org.apache.flink.table.catalog.exceptions.PartitionNotExistException;
 import org.apache.flink.table.catalog.exceptions.TableNotExistException;
 import org.apache.flink.table.catalog.hive.HiveCatalog;
+import org.apache.flink.table.catalog.hive.client.HiveShim;
+import org.apache.flink.table.catalog.hive.factories.HiveFunctionDefinitionFactory;
+import org.apache.flink.table.catalog.hive.util.HiveTableUtil;
+import org.apache.flink.table.catalog.hive.util.HiveTypeUtil;
+import org.apache.flink.table.factories.FactoryUtil;
+import org.apache.flink.table.functions.FunctionDefinition;
+import org.apache.flink.table.operations.CatalogSinkModifyOperation;
+import org.apache.flink.table.operations.DescribeTableOperation;
+import org.apache.flink.table.operations.Operation;
+import org.apache.flink.table.operations.ShowDatabasesOperation;
+import org.apache.flink.table.operations.ShowFunctionsOperation;
+import org.apache.flink.table.operations.ShowPartitionsOperation;
+import org.apache.flink.table.operations.ShowTablesOperation;
+import org.apache.flink.table.operations.ShowViewsOperation;
+import org.apache.flink.table.operations.UseDatabaseOperation;
+import org.apache.flink.table.operations.ddl.AddPartitionsOperation;
+import org.apache.flink.table.operations.ddl.AlterDatabaseOperation;
+import org.apache.flink.table.operations.ddl.AlterPartitionPropertiesOperation;
+import org.apache.flink.table.operations.ddl.AlterTableOptionsOperation;
+import org.apache.flink.table.operations.ddl.AlterTableRenameOperation;
+import org.apache.flink.table.operations.ddl.AlterTableSchemaOperation;
+import org.apache.flink.table.operations.ddl.AlterViewAsOperation;
+import org.apache.flink.table.operations.ddl.AlterViewPropertiesOperation;
+import org.apache.flink.table.operations.ddl.AlterViewRenameOperation;
+import org.apache.flink.table.operations.ddl.CreateCatalogFunctionOperation;
+import org.apache.flink.table.operations.ddl.CreateDatabaseOperation;
+import org.apache.flink.table.operations.ddl.CreateTableASOperation;
+import org.apache.flink.table.operations.ddl.CreateTableOperation;
+import org.apache.flink.table.operations.ddl.CreateTempSystemFunctionOperation;
+import org.apache.flink.table.operations.ddl.CreateViewOperation;
+import org.apache.flink.table.operations.ddl.DropCatalogFunctionOperation;
+import org.apache.flink.table.operations.ddl.DropDatabaseOperation;
+import org.apache.flink.table.operations.ddl.DropPartitionsOperation;
+import org.apache.flink.table.operations.ddl.DropTableOperation;
+import org.apache.flink.table.operations.ddl.DropTempSystemFunctionOperation;
+import org.apache.flink.table.operations.ddl.DropViewOperation;
+import org.apache.flink.table.planner.delegation.hive.HiveParser;
+import org.apache.flink.table.planner.delegation.hive.HiveParserCalcitePlanner;
 import org.apache.flink.table.planner.delegation.hive.HiveParserConstants;
+import org.apache.flink.table.planner.delegation.hive.HiveParserDMLHelper;
 import org.apache.flink.table.planner.delegation.hive.copy.HiveASTParseUtils;
 import org.apache.flink.table.planner.delegation.hive.copy.HiveParserASTNode;
 import org.apache.flink.table.planner.delegation.hive.copy.HiveParserAuthorizationParseUtils;
 import org.apache.flink.table.planner.delegation.hive.copy.HiveParserBaseSemanticAnalyzer;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserContext;
 import org.apache.flink.table.planner.delegation.hive.copy.HiveParserQueryState;
 import org.apache.flink.table.planner.delegation.hive.copy.HiveParserStorageFormat;
-import org.apache.flink.table.planner.delegation.hive.desc.CreateTableASDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.DropPartitionDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserAlterDatabaseDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserAlterTableDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserCreateTableDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserCreateTableDesc.NotNullConstraint;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserCreateTableDesc.PrimaryKey;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserCreateViewDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserDropDatabaseDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserDropFunctionDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserDropTableDesc;
-import org.apache.flink.table.planner.delegation.hive.desc.HiveParserShowTablesDesc;
+import org.apache.flink.table.planner.utils.OperationConverterUtils;
 
 import org.antlr.runtime.tree.CommonTree;
-import org.apache.hadoop.fs.Path;
+import org.apache.calcite.rel.RelNode;
 import org.apache.hadoop.hive.common.StatsSetupConst;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.TableType;
@@ -54,36 +107,20 @@ import org.apache.hadoop.hive.metastore.api.SkewedInfo;
 import org.apache.hadoop.hive.ql.ErrorMsg;
 import org.apache.hadoop.hive.ql.exec.FunctionUtils;
 import org.apache.hadoop.hive.ql.exec.Utilities;
-import org.apache.hadoop.hive.ql.hooks.ReadEntity;
-import org.apache.hadoop.hive.ql.hooks.WriteEntity;
 import org.apache.hadoop.hive.ql.metadata.Table;
 import org.apache.hadoop.hive.ql.parse.EximUtil;
 import org.apache.hadoop.hive.ql.parse.SemanticException;
-import org.apache.hadoop.hive.ql.plan.AddPartitionDesc;
 import org.apache.hadoop.hive.ql.plan.AlterTableDesc;
-import org.apache.hadoop.hive.ql.plan.CreateDatabaseDesc;
-import org.apache.hadoop.hive.ql.plan.CreateFunctionDesc;
-import org.apache.hadoop.hive.ql.plan.DDLWork;
-import org.apache.hadoop.hive.ql.plan.DescDatabaseDesc;
-import org.apache.hadoop.hive.ql.plan.DescFunctionDesc;
-import org.apache.hadoop.hive.ql.plan.DescTableDesc;
-import org.apache.hadoop.hive.ql.plan.DropFunctionDesc;
-import org.apache.hadoop.hive.ql.plan.FunctionWork;
 import org.apache.hadoop.hive.ql.plan.HiveOperation;
 import org.apache.hadoop.hive.ql.plan.PrincipalDesc;
-import org.apache.hadoop.hive.ql.plan.ShowDatabasesDesc;
-import org.apache.hadoop.hive.ql.plan.ShowFunctionsDesc;
-import org.apache.hadoop.hive.ql.plan.ShowPartitionsDesc;
-import org.apache.hadoop.hive.ql.plan.SwitchDatabaseDesc;
-import org.apache.hadoop.hive.ql.session.SessionState;
 import org.apache.hadoop.hive.serde.serdeConstants;
 import org.apache.hadoop.hive.serde2.typeinfo.CharTypeInfo;
 import org.apache.hadoop.hive.serde2.typeinfo.DecimalTypeInfo;
+import org.apache.hadoop.hive.serde2.typeinfo.TypeInfoUtils;
 import org.apache.hadoop.hive.serde2.typeinfo.VarcharTypeInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -91,11 +128,46 @@ import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
+
+import static org.apache.flink.sql.parser.hive.ddl.HiveDDLUtils.COL_DELIMITER;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabase.ALTER_DATABASE_OP;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabaseOwner.DATABASE_OWNER_NAME;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabaseOwner.DATABASE_OWNER_TYPE;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.ALTER_COL_CASCADE;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.ALTER_TABLE_OP;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.AlterTableOp.ALTER_COLUMNS;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.AlterTableOp.CHANGE_FILE_FORMAT;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.AlterTableOp.CHANGE_LOCATION;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.AlterTableOp.CHANGE_SERDE_PROPS;
+import static org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveTable.AlterTableOp.CHANGE_TBL_PROPS;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveDatabase.DATABASE_LOCATION_URI;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.COLLECTION_DELIM;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.ESCAPE_CHAR;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.FIELD_DELIM;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.LINE_DELIM;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.MAPKEY_DELIM;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.SERDE_INFO_PROP_PREFIX;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.SERDE_LIB_CLASS_NAME;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableRowFormat.SERIALIZATION_NULL_FORMAT;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableStoredAs.STORED_AS_FILE_FORMAT;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableStoredAs.STORED_AS_INPUT_FORMAT;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.HiveTableStoredAs.STORED_AS_OUTPUT_FORMAT;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.NOT_NULL_COLS;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.NOT_NULL_CONSTRAINT_TRAITS;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.PK_CONSTRAINT_TRAIT;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.TABLE_IS_EXTERNAL;
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveTable.TABLE_LOCATION_URI;
+import static org.apache.flink.table.planner.delegation.hive.copy.HiveParserBaseSemanticAnalyzer.HiveParserRowFormatParams;
+import static org.apache.flink.table.planner.delegation.hive.copy.HiveParserBaseSemanticAnalyzer.NotNullConstraint;
+import static org.apache.flink.table.planner.delegation.hive.copy.HiveParserBaseSemanticAnalyzer.PrimaryKey;
 
 /**
  * Ported hive's org.apache.hadoop.hive.ql.parse.DDLSemanticAnalyzer, and also incorporated
- * functionalities from SemanticAnalyzer and FunctionSemanticAnalyzer.
+ * functionalities from SemanticAnalyzer and FunctionSemanticAnalyzer. It's mainly used to convert
+ * {@link HiveParserASTNode} to the corresponding {@link Operation}.
  */
 public class HiveParserDDLSemanticAnalyzer {
     private static final Logger LOG = LoggerFactory.getLogger(HiveParserDDLSemanticAnalyzer.class);
@@ -105,7 +177,13 @@ public class HiveParserDDLSemanticAnalyzer {
     private final HiveConf conf;
     private final HiveParserQueryState queryState;
     private final HiveCatalog hiveCatalog;
+    private final CatalogManager catalogManager;
     private final String currentDB;
+    private final HiveParser hiveParser;
+    private final HiveFunctionDefinitionFactory funcDefFactory;
+    private final HiveShim hiveShim;
+    private final HiveParserContext context;
+    private final HiveParserDMLHelper dmlHelper;
 
     static {
         TokenToTypeName.put(HiveASTParser.TOK_BOOLEAN, serdeConstants.BOOLEAN_TYPE_NAME);
@@ -160,12 +238,24 @@ public class HiveParserDDLSemanticAnalyzer {
     }
 
     public HiveParserDDLSemanticAnalyzer(
-            HiveParserQueryState queryState, HiveCatalog hiveCatalog, String currentDB)
+            HiveParserQueryState queryState,
+            HiveCatalog hiveCatalog,
+            CatalogManager catalogManager,
+            HiveParser hiveParser,
+            HiveShim hiveShim,
+            HiveParserContext context,
+            HiveParserDMLHelper dmlHelper)
             throws SemanticException {
         this.queryState = queryState;
         this.conf = queryState.getConf();
         this.hiveCatalog = hiveCatalog;
-        this.currentDB = currentDB;
+        this.currentDB = catalogManager.getCurrentDatabase();
+        this.catalogManager = catalogManager;
+        this.hiveParser = hiveParser;
+        this.funcDefFactory = new HiveFunctionDefinitionFactory(hiveShim);
+        this.hiveShim = hiveShim;
+        this.context = context;
+        this.dmlHelper = dmlHelper;
         reservedPartitionValues = new HashSet<>();
         // Partition can't have this name
         reservedPartitionValues.add(HiveConf.getVar(conf, HiveConf.ConfVars.DEFAULTPARTITIONNAME));
@@ -197,18 +287,9 @@ public class HiveParserDDLSemanticAnalyzer {
         return new ObjectPath(parts[0], parts[1]);
     }
 
-    private HashSet<ReadEntity> getInputs() {
-        return new HashSet<>();
-    }
-
-    private HashSet<WriteEntity> getOutputs() {
-        return new HashSet<>();
-    }
-
-    public Serializable analyzeInternal(HiveParserASTNode input) throws SemanticException {
-
+    public Operation convertToOperation(HiveParserASTNode input) throws SemanticException {
         HiveParserASTNode ast = input;
-        Serializable res = null;
+        Operation res = null;
         switch (ast.getType()) {
             case HiveASTParser.TOK_ALTERTABLE:
                 {
@@ -222,9 +303,8 @@ public class HiveParserDDLSemanticAnalyzer {
                     if (partSpecNode != null) {
                         partSpec = getPartSpec(partSpecNode);
                     }
-
                     if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_RENAME) {
-                        res = analyzeAlterTableRename(qualified, ast, false);
+                        res = convertAlterTableRename(qualified, ast, false);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_TOUCH) {
                         handleUnsupportedOperation(ast);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_ARCHIVE) {
@@ -232,41 +312,41 @@ public class HiveParserDDLSemanticAnalyzer {
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_UNARCHIVE) {
                         handleUnsupportedOperation(ast);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_ADDCOLS) {
-                        res = analyzeAlterTableModifyCols(qualified, ast, partSpec, false);
+                        res = convertAlterTableModifyCols(qualified, ast, false);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_REPLACECOLS) {
-                        res = analyzeAlterTableModifyCols(qualified, ast, partSpec, true);
+                        res = convertAlterTableModifyCols(qualified, ast, true);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_RENAMECOL) {
-                        res = analyzeAlterTableRenameCol(qualified, ast, partSpec);
+                        res = convertAlterTableRenameCol(qualified, ast);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_ADDPARTS) {
-                        res = analyzeAlterTableAddParts(qualified, ast, false);
+                        res = convertAlterTableAddParts(qualified, ast, false);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_DROPPARTS) {
-                        res = analyzeAlterTableDropParts(qualified, ast, false);
+                        res = convertAlterTableDropParts(qualified, ast, false);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_PARTCOLTYPE) {
                         handleUnsupportedOperation(ast);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_PROPERTIES) {
-                        res = analyzeAlterTableProps(qualified, null, ast, false, false);
+                        res = convertAlterTableProps(qualified, null, ast, false, false);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_DROPPROPERTIES) {
-                        res = analyzeAlterTableProps(qualified, null, ast, false, true);
+                        res = convertAlterTableProps(qualified, null, ast, false, true);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_UPDATESTATS) {
-                        res = analyzeAlterTableProps(qualified, partSpec, ast, false, false);
+                        res = convertAlterTableProps(qualified, partSpec, ast, false, false);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_SKEWED) {
                         handleUnsupportedOperation(ast);
                     } else if (ast.getType() == HiveASTParser.TOK_ALTERTABLE_EXCHANGEPARTITION) {
                         handleUnsupportedOperation(ast);
                     } else if (ast.getToken().getType()
                             == HiveASTParser.TOK_ALTERTABLE_FILEFORMAT) {
-                        res = analyzeAlterTableFileFormat(ast, tableName, partSpec);
+                        res = convertAlterTableFileFormat(ast, tableName, partSpec);
                     } else if (ast.getToken().getType() == HiveASTParser.TOK_ALTERTABLE_LOCATION) {
-                        res = analyzeAlterTableLocation(ast, tableName, partSpec);
+                        res = convertAlterTableLocation(ast, tableName, partSpec);
                     } else if (ast.getToken().getType()
                             == HiveASTParser.TOK_ALTERTABLE_MERGEFILES) {
                         handleUnsupportedOperation(ast);
                     } else if (ast.getToken().getType()
                             == HiveASTParser.TOK_ALTERTABLE_SERIALIZER) {
-                        res = analyzeAlterTableSerde(ast, tableName, partSpec);
+                        res = convertAlterTableSerde(ast, tableName, partSpec);
                     } else if (ast.getToken().getType()
                             == HiveASTParser.TOK_ALTERTABLE_SERDEPROPERTIES) {
-                        res = analyzeAlterTableSerdeProps(ast, tableName, partSpec);
+                        res = convertAlterTableSerdeProps(ast, tableName, partSpec);
                     } else if (ast.getToken().getType()
                             == HiveASTParser.TOK_ALTERTABLE_RENAMEPART) {
                         handleUnsupportedOperation(ast);
@@ -295,52 +375,46 @@ public class HiveParserDDLSemanticAnalyzer {
                     break;
                 }
             case HiveASTParser.TOK_DROPTABLE:
-                res = analyzeDropTable(ast, null);
+                res = convertDropTable(ast, null);
                 break;
             case HiveASTParser.TOK_DESCTABLE:
-                res = analyzeDescribeTable(ast);
+                res = convertDescribeTable(ast);
                 break;
             case HiveASTParser.TOK_SHOWDATABASES:
-                res = analyzeShowDatabases(ast);
+                res = convertShowDatabases();
                 break;
             case HiveASTParser.TOK_SHOWTABLES:
-                res = analyzeShowTables(ast, false);
+                res = convertShowTables(ast, false);
                 break;
             case HiveASTParser.TOK_SHOWFUNCTIONS:
-                res = analyzeShowFunctions(ast);
+                res = convertShowFunctions(ast);
                 break;
             case HiveASTParser.TOK_SHOWVIEWS:
-                res = analyzeShowTables(ast, true);
-                break;
-            case HiveASTParser.TOK_DESCFUNCTION:
-                res = analyzeDescFunction(ast);
-                break;
-            case HiveASTParser.TOK_DESCDATABASE:
-                res = analyzeDescDatabase(ast);
+                res = convertShowTables(ast, true);
                 break;
             case HiveASTParser.TOK_DROPVIEW:
-                res = analyzeDropTable(ast, TableType.VIRTUAL_VIEW);
+                res = convertDropTable(ast, TableType.VIRTUAL_VIEW);
                 break;
             case HiveASTParser.TOK_ALTERVIEW:
                 {
                     if (ast.getChild(1).getType() == HiveASTParser.TOK_QUERY) {
                         // alter view as
-                        res = analyzeCreateView(ast);
+                        res = convertCreateView(ast);
                     } else {
                         String[] qualified =
                                 HiveParserBaseSemanticAnalyzer.getQualifiedTableName(
                                         (HiveParserASTNode) ast.getChild(0));
                         ast = (HiveParserASTNode) ast.getChild(1);
                         if (ast.getType() == HiveASTParser.TOK_ALTERVIEW_PROPERTIES) {
-                            res = analyzeAlterTableProps(qualified, null, ast, true, false);
+                            res = convertAlterTableProps(qualified, null, ast, true, false);
                         } else if (ast.getType() == HiveASTParser.TOK_ALTERVIEW_DROPPROPERTIES) {
-                            res = analyzeAlterTableProps(qualified, null, ast, true, true);
+                            res = convertAlterTableProps(qualified, null, ast, true, true);
                         } else if (ast.getType() == HiveASTParser.TOK_ALTERVIEW_ADDPARTS) {
                             handleUnsupportedOperation("ADD PARTITION for view is not supported");
                         } else if (ast.getType() == HiveASTParser.TOK_ALTERVIEW_DROPPARTS) {
                             handleUnsupportedOperation("DROP PARTITION for view is not supported");
                         } else if (ast.getType() == HiveASTParser.TOK_ALTERVIEW_RENAME) {
-                            res = analyzeAlterTableRename(qualified, ast, true);
+                            res = convertAlterTableRename(qualified, ast, true);
                         } else {
                             throw new ValidationException(
                                     "Unknown AST node for ALTER VIEW: " + ast);
@@ -349,38 +423,40 @@ public class HiveParserDDLSemanticAnalyzer {
                     break;
                 }
             case HiveASTParser.TOK_SHOWPARTITIONS:
-                res = analyzeShowPartitions(ast);
+                res = convertShowPartitions(ast);
                 break;
             case HiveASTParser.TOK_CREATEDATABASE:
-                res = analyzeCreateDatabase(ast);
+                res = convertCreateDatabase(ast);
                 break;
             case HiveASTParser.TOK_DROPDATABASE:
-                res = analyzeDropDatabase(ast);
+                res = convertDropDatabase(ast);
                 break;
             case HiveASTParser.TOK_SWITCHDATABASE:
-                res = analyzeSwitchDatabase(ast);
+                res = convertSwitchDatabase(ast);
                 break;
             case HiveASTParser.TOK_ALTERDATABASE_PROPERTIES:
-                res = analyzeAlterDatabaseProperties(ast);
+                res = convertAlterDatabaseProperties(ast);
                 break;
             case HiveASTParser.TOK_ALTERDATABASE_OWNER:
-                res = analyzeAlterDatabaseOwner(ast);
+                res = convertAlterDatabaseOwner(ast);
                 break;
             case HiveASTParser.TOK_ALTERDATABASE_LOCATION:
-                res = analyzeAlterDatabaseLocation(ast);
+                res = convertAlterDatabaseLocation(ast);
                 break;
             case HiveASTParser.TOK_CREATETABLE:
-                res = analyzeCreateTable(ast);
+                res = convertCreateTable(ast);
                 break;
             case HiveASTParser.TOK_CREATEVIEW:
-                res = analyzeCreateView(ast);
+                res = convertCreateView(ast);
                 break;
             case HiveASTParser.TOK_CREATEFUNCTION:
-                res = analyzerCreateFunction(ast);
+                res = convertCreateFunction(ast);
                 break;
             case HiveASTParser.TOK_DROPFUNCTION:
-                res = analyzeDropFunction(ast);
+                res = convertDropFunction(ast);
                 break;
+            case HiveASTParser.TOK_DESCFUNCTION:
+            case HiveASTParser.TOK_DESCDATABASE:
             case HiveASTParser.TOK_TRUNCATETABLE:
             case HiveASTParser.TOK_CREATEINDEX:
             case HiveASTParser.TOK_DROPINDEX:
@@ -422,20 +498,22 @@ public class HiveParserDDLSemanticAnalyzer {
         return res;
     }
 
-    private Serializable analyzeDropFunction(HiveParserASTNode ast) {
+    private Operation convertDropFunction(HiveParserASTNode ast) {
         // ^(TOK_DROPFUNCTION identifier ifExists? $temp?)
         String functionName = ast.getChild(0).getText();
         boolean ifExists = (ast.getFirstChildWithType(HiveASTParser.TOK_IFEXISTS) != null);
 
         boolean isTemporaryFunction =
                 (ast.getFirstChildWithType(HiveASTParser.TOK_TEMPORARY) != null);
-        DropFunctionDesc desc = new DropFunctionDesc();
-        desc.setFunctionName(functionName);
-        desc.setTemp(isTemporaryFunction);
-        return new HiveParserDropFunctionDesc(desc, ifExists);
+        if (isTemporaryFunction) {
+            return new DropTempSystemFunctionOperation(functionName, ifExists);
+        } else {
+            ObjectIdentifier identifier = parseObjectIdentifier(functionName);
+            return new DropCatalogFunctionOperation(identifier, ifExists, false);
+        }
     }
 
-    private Serializable analyzerCreateFunction(HiveParserASTNode ast) {
+    private Operation convertCreateFunction(HiveParserASTNode ast) {
         // ^(TOK_CREATEFUNCTION identifier StringLiteral ({isTempFunction}? => TOK_TEMPORARY))
         String functionName = ast.getChild(0).getText().toLowerCase();
         boolean isTemporaryFunction =
@@ -449,15 +527,24 @@ public class HiveParserDDLSemanticAnalyzer {
                     "Temporary function cannot be created with a qualified name.");
         }
 
-        CreateFunctionDesc desc = new CreateFunctionDesc();
-        desc.setFunctionName(functionName);
-        desc.setTemp(isTemporaryFunction);
-        desc.setClassName(className);
-        desc.setResources(Collections.emptyList());
-        return new FunctionWork(desc);
+        if (isTemporaryFunction) {
+            // hive's temporary function is more like flink's temp system function, e.g. doesn't
+            // belong to a catalog/db
+            // the DDL analyzer makes sure temp function name is not a compound one
+            FunctionDefinition funcDefinition =
+                    funcDefFactory.createFunctionDefinition(
+                            functionName,
+                            new CatalogFunctionImpl(className, FunctionLanguage.JAVA));
+            return new CreateTempSystemFunctionOperation(functionName, false, funcDefinition);
+        } else {
+            ObjectIdentifier identifier = parseObjectIdentifier(functionName);
+            CatalogFunction catalogFunction =
+                    new CatalogFunctionImpl(className, FunctionLanguage.JAVA);
+            return new CreateCatalogFunctionOperation(identifier, catalogFunction, false, false);
+        }
     }
 
-    private Serializable analyzeCreateView(HiveParserASTNode ast) throws SemanticException {
+    private Operation convertCreateView(HiveParserASTNode ast) throws SemanticException {
         String[] qualTabName =
                 HiveParserBaseSemanticAnalyzer.getQualifiedTableName(
                         (HiveParserASTNode) ast.getChild(0));
@@ -473,8 +560,6 @@ public class HiveParserDDLSemanticAnalyzer {
         if (isMaterialized) {
             handleUnsupportedOperation("MATERIALIZED VIEW is not supported");
         }
-        HiveParserBaseSemanticAnalyzer.HiveParserRowFormatParams rowFormatParams =
-                new HiveParserBaseSemanticAnalyzer.HiveParserRowFormatParams();
         HiveParserStorageFormat storageFormat = new HiveParserStorageFormat(conf);
 
         LOG.info("Creating view " + dbDotTable + " position=" + ast.getCharPositionInLine());
@@ -532,11 +617,46 @@ public class HiveParserDDLSemanticAnalyzer {
         }
 
         queryState.setCommandType(HiveOperation.CREATEVIEW);
-        return new HiveParserCreateViewDesc(
-                dbDotTable, cols, comment, tblProps, ifNotExists, isAlterViewAs, selectStmt);
+
+        HiveParserCreateViewInfo createViewInfo =
+                new HiveParserCreateViewInfo(dbDotTable, cols, selectStmt);
+        hiveParser.analyzeCreateView(createViewInfo, context, queryState, hiveShim);
+
+        ObjectIdentifier viewIdentifier = parseObjectIdentifier(createViewInfo.getCompoundName());
+        TableSchema schema =
+                HiveTableUtil.createTableSchema(
+                        createViewInfo.getSchema(),
+                        Collections.emptyList(),
+                        Collections.emptySet(),
+                        null);
+        Map<String, String> props = new HashMap<>();
+        if (isAlterViewAs) {
+            CatalogBaseTable baseTable = getCatalogBaseTable(viewIdentifier);
+            if (baseTable instanceof CatalogTable) {
+                throw new ValidationException("ALTER VIEW for a table is not allowed");
+            }
+            props.putAll(baseTable.getOptions());
+            comment = baseTable.getComment();
+        } else {
+            if (tblProps != null) {
+                props.putAll(tblProps);
+            }
+        }
+        CatalogView catalogView =
+                new CatalogViewImpl(
+                        createViewInfo.getOriginalText(),
+                        createViewInfo.getExpandedText(),
+                        schema,
+                        props,
+                        comment);
+        if (isAlterViewAs) {
+            return new AlterViewAsOperation(viewIdentifier, catalogView);
+        } else {
+            return new CreateViewOperation(viewIdentifier, catalogView, ifNotExists, false);
+        }
     }
 
-    private Serializable analyzeCreateTable(HiveParserASTNode ast) throws SemanticException {
+    private Operation convertCreateTable(HiveParserASTNode ast) throws SemanticException {
         String[] qualifiedTabName =
                 HiveParserBaseSemanticAnalyzer.getQualifiedTableName(
                         (HiveParserASTNode) ast.getChild(0));
@@ -694,7 +814,7 @@ public class HiveParserDDLSemanticAnalyzer {
         switch (commandType) {
             case createTable: // REGULAR CREATE TABLE DDL
                 tblProps = addDefaultProperties(tblProps);
-                return new HiveParserCreateTableDesc(
+                return convertCreateTable(
                         dbDotTab,
                         isExt,
                         ifNotExists,
@@ -716,29 +836,188 @@ public class HiveParserDDLSemanticAnalyzer {
             case ctas: // create table as select
                 tblProps = addDefaultProperties(tblProps);
 
-                HiveParserCreateTableDesc createTableDesc =
-                        new HiveParserCreateTableDesc(
-                                dbDotTab,
-                                isExt,
-                                ifNotExists,
-                                isTemporary,
-                                cols,
-                                partCols,
-                                comment,
-                                location,
-                                tblProps,
-                                rowFormatParams,
-                                storageFormat,
-                                primaryKeys,
-                                notNulls);
-                return new CreateTableASDesc(createTableDesc, selectStmt);
+                // analyze the query
+                HiveParserCalcitePlanner calcitePlanner =
+                        hiveParser.createCalcitePlanner(context, queryState, hiveShim);
+                calcitePlanner.setCtasCols(cols);
+                RelNode queryRelNode = calcitePlanner.genLogicalPlan(selectStmt);
+                // create a table to represent the dest table
+                String[] dbTblName = dbDotTab.split("\\.");
+                Table destTable = new Table(Table.getEmptyTable(dbTblName[0], dbTblName[1]));
+                destTable.getSd().setCols(cols);
+                // create the insert operation
+                CatalogSinkModifyOperation insertOperation =
+                        dmlHelper.createInsertOperation(
+                                queryRelNode,
+                                destTable,
+                                Collections.emptyMap(),
+                                Collections.emptyList(),
+                                false);
+
+                CreateTableOperation createTableOperation =
+                        (CreateTableOperation)
+                                convertCreateTable(
+                                        dbDotTab,
+                                        isExt,
+                                        ifNotExists,
+                                        isTemporary,
+                                        cols,
+                                        partCols,
+                                        comment,
+                                        location,
+                                        tblProps,
+                                        rowFormatParams,
+                                        storageFormat,
+                                        primaryKeys,
+                                        notNulls);
+
+                return new CreateTableASOperation(createTableOperation, insertOperation);
             default:
                 throw new ValidationException("Unrecognized command.");
         }
     }
 
-    private Serializable analyzeAlterDatabaseProperties(HiveParserASTNode ast) {
+    private Operation convertCreateTable(
+            String compoundName,
+            boolean isExternal,
+            boolean ifNotExists,
+            boolean isTemporary,
+            List<FieldSchema> cols,
+            List<FieldSchema> partCols,
+            String comment,
+            String location,
+            Map<String, String> tblProps,
+            HiveParserRowFormatParams rowFormatParams,
+            HiveParserStorageFormat storageFormat,
+            List<PrimaryKey> primaryKeys,
+            List<NotNullConstraint> notNullConstraints) {
+        Map<String, String> props = new HashMap<>();
+        if (tblProps != null) {
+            props.putAll(tblProps);
+        }
+        markHiveConnector(props);
+        // external
+        if (isExternal) {
+            props.put(TABLE_IS_EXTERNAL, "true");
+        }
+        // PK trait
+        UniqueConstraint uniqueConstraint = null;
+        if (primaryKeys != null && !primaryKeys.isEmpty()) {
+            PrimaryKey primaryKey = primaryKeys.get(0);
+            byte trait = 0;
+            if (primaryKey.isEnable()) {
+                trait = HiveDDLUtils.enableConstraint(trait);
+            }
+            if (primaryKey.isValidate()) {
+                trait = HiveDDLUtils.validateConstraint(trait);
+            }
+            if (primaryKey.isRely()) {
+                trait = HiveDDLUtils.relyConstraint(trait);
+            }
+            props.put(PK_CONSTRAINT_TRAIT, String.valueOf(trait));
+            List<String> pkCols =
+                    primaryKeys.stream().map(PrimaryKey::getPk).collect(Collectors.toList());
+            String constraintName = primaryKey.getConstraintName();
+            if (constraintName == null) {
+                constraintName = pkCols.stream().collect(Collectors.joining("_", "PK_", ""));
+            }
+            uniqueConstraint = UniqueConstraint.primaryKey(constraintName, pkCols);
+        }
+        // NOT NULL constraints
+        List<String> notNullCols = new ArrayList<>();
+        if (!notNullConstraints.isEmpty()) {
+            List<String> traits = new ArrayList<>();
+            for (NotNullConstraint notNull : notNullConstraints) {
+                byte trait = 0;
+                if (notNull.isEnable()) {
+                    trait = HiveDDLUtils.enableConstraint(trait);
+                }
+                if (notNull.isValidate()) {
+                    trait = HiveDDLUtils.validateConstraint(trait);
+                }
+                if (notNull.isRely()) {
+                    trait = HiveDDLUtils.relyConstraint(trait);
+                }
+                traits.add(String.valueOf(trait));
+                notNullCols.add(notNull.getColName());
+            }
+            props.put(NOT_NULL_CONSTRAINT_TRAITS, String.join(COL_DELIMITER, traits));
+            props.put(NOT_NULL_COLS, String.join(COL_DELIMITER, notNullCols));
+        }
+        // row format
+        if (rowFormatParams != null) {
+            encodeRowFormat(rowFormatParams, props);
+        }
+        // storage format
+        if (storageFormat != null) {
+            encodeStorageFormat(storageFormat, props);
+        }
+        // location
+        if (location != null) {
+            props.put(TABLE_LOCATION_URI, location);
+        }
+        ObjectIdentifier identifier = parseObjectIdentifier(compoundName);
+        Set<String> notNullColSet = new HashSet<>(notNullCols);
+        if (uniqueConstraint != null) {
+            notNullColSet.addAll(uniqueConstraint.getColumns());
+        }
+        TableSchema tableSchema =
+                HiveTableUtil.createTableSchema(cols, partCols, notNullColSet, uniqueConstraint);
+        return new CreateTableOperation(
+                identifier,
+                new CatalogTableImpl(
+                        tableSchema, HiveCatalog.getFieldNames(partCols), props, comment),
+                ifNotExists,
+                isTemporary);
+    }
 
+    private void markHiveConnector(Map<String, String> props) {
+        props.put(FactoryUtil.CONNECTOR.key(), SqlCreateHiveTable.IDENTIFIER);
+    }
+
+    private void encodeRowFormat(
+            HiveParserRowFormatParams rowFormatParams, Map<String, String> props) {
+        if (rowFormatParams.getFieldDelim() != null) {
+            props.put(FIELD_DELIM, rowFormatParams.getFieldDelim());
+        }
+        if (rowFormatParams.getCollItemDelim() != null) {
+            props.put(COLLECTION_DELIM, rowFormatParams.getCollItemDelim());
+        }
+        if (rowFormatParams.getMapKeyDelim() != null) {
+            props.put(MAPKEY_DELIM, rowFormatParams.getMapKeyDelim());
+        }
+        if (rowFormatParams.getFieldEscape() != null) {
+            props.put(ESCAPE_CHAR, rowFormatParams.getFieldEscape());
+        }
+        if (rowFormatParams.getLineDelim() != null) {
+            props.put(LINE_DELIM, rowFormatParams.getLineDelim());
+        }
+        if (rowFormatParams.getNullFormat() != null) {
+            props.put(SERIALIZATION_NULL_FORMAT, rowFormatParams.getNullFormat());
+        }
+    }
+
+    private void encodeStorageFormat(
+            HiveParserStorageFormat storageFormat, Map<String, String> props) {
+        String serdeName = storageFormat.getSerde();
+        if (serdeName != null) {
+            props.put(SERDE_LIB_CLASS_NAME, serdeName);
+        }
+        Map<String, String> serdeProps = storageFormat.getSerdeProps();
+        if (serdeProps != null) {
+            for (String serdeKey : serdeProps.keySet()) {
+                props.put(SERDE_INFO_PROP_PREFIX + serdeKey, serdeProps.get(serdeKey));
+            }
+        }
+        if (storageFormat.getInputFormat() != null) {
+            props.put(STORED_AS_INPUT_FORMAT, storageFormat.getInputFormat());
+        }
+        if (storageFormat.getOutputFormat() != null) {
+            props.put(STORED_AS_OUTPUT_FORMAT, storageFormat.getOutputFormat());
+        }
+    }
+
+    private Operation convertAlterDatabaseProperties(HiveParserASTNode ast) {
         String dbName =
                 HiveParserBaseSemanticAnalyzer.unescapeIdentifier(ast.getChild(0).getText());
         Map<String, String> dbProps = null;
@@ -754,10 +1033,15 @@ public class HiveParserDDLSemanticAnalyzer {
                             "Unknown AST node for ALTER DATABASE PROPERTIES: " + childNode);
             }
         }
-        return HiveParserAlterDatabaseDesc.alterProps(dbName, dbProps);
+        CatalogDatabase originDB = getDatabase(dbName);
+        Map<String, String> props = new HashMap<>(originDB.getProperties());
+        props.put(ALTER_DATABASE_OP, SqlAlterHiveDatabase.AlterHiveDatabaseOp.CHANGE_PROPS.name());
+        props.putAll(dbProps);
+        CatalogDatabase newDB = new CatalogDatabaseImpl(props, originDB.getComment());
+        return new AlterDatabaseOperation(catalogManager.getCurrentCatalog(), dbName, newDB);
     }
 
-    private Serializable analyzeAlterDatabaseOwner(HiveParserASTNode ast) {
+    private Operation convertAlterDatabaseOwner(HiveParserASTNode ast) {
         String dbName =
                 HiveParserBaseSemanticAnalyzer.getUnescapedName(
                         (HiveParserASTNode) ast.getChild(0));
@@ -773,20 +1057,31 @@ public class HiveParserDDLSemanticAnalyzer {
         if (principalDesc.getType() == null) {
             throw new ValidationException("Owner type " + nullCmdMsg);
         }
-
-        return HiveParserAlterDatabaseDesc.alterOwner(dbName, principalDesc);
+        CatalogDatabase originDB = getDatabase(dbName);
+        Map<String, String> props = new HashMap<>(originDB.getProperties());
+        props.put(ALTER_DATABASE_OP, SqlAlterHiveDatabase.AlterHiveDatabaseOp.CHANGE_OWNER.name());
+        props.put(DATABASE_OWNER_NAME, principalDesc.getName());
+        props.put(DATABASE_OWNER_TYPE, principalDesc.getType().name().toLowerCase());
+        CatalogDatabase newDB = new CatalogDatabaseImpl(props, originDB.getComment());
+        return new AlterDatabaseOperation(catalogManager.getCurrentCatalog(), dbName, newDB);
     }
 
-    private Serializable analyzeAlterDatabaseLocation(HiveParserASTNode ast) {
+    private Operation convertAlterDatabaseLocation(HiveParserASTNode ast) {
         String dbName =
                 HiveParserBaseSemanticAnalyzer.getUnescapedName(
                         (HiveParserASTNode) ast.getChild(0));
         String newLocation =
                 HiveParserBaseSemanticAnalyzer.unescapeSQLString(ast.getChild(1).getText());
-        return HiveParserAlterDatabaseDesc.alterLocation(dbName, newLocation);
+        CatalogDatabase originDB = getDatabase(dbName);
+        Map<String, String> props = new HashMap<>(originDB.getProperties());
+        props.put(
+                ALTER_DATABASE_OP, SqlAlterHiveDatabase.AlterHiveDatabaseOp.CHANGE_LOCATION.name());
+        props.put(DATABASE_LOCATION_URI, newLocation);
+        CatalogDatabase newDB = new CatalogDatabaseImpl(props, originDB.getComment());
+        return new AlterDatabaseOperation(catalogManager.getCurrentCatalog(), dbName, newDB);
     }
 
-    private Serializable analyzeCreateDatabase(HiveParserASTNode ast) {
+    private Operation convertCreateDatabase(HiveParserASTNode ast) {
         String dbName =
                 HiveParserBaseSemanticAnalyzer.unescapeIdentifier(ast.getChild(0).getText());
         boolean ifNotExists = false;
@@ -819,15 +1114,20 @@ public class HiveParserDDLSemanticAnalyzer {
             }
         }
 
-        CreateDatabaseDesc createDatabaseDesc =
-                new CreateDatabaseDesc(dbName, dbComment, dbLocation, ifNotExists);
+        Map<String, String> props = new HashMap<>();
         if (dbProps != null) {
-            createDatabaseDesc.setDatabaseProperties(dbProps);
+            props.putAll(dbProps);
         }
-        return new DDLWork(getInputs(), getOutputs(), createDatabaseDesc);
+
+        if (dbLocation != null) {
+            props.put(DATABASE_LOCATION_URI, dbLocation);
+        }
+        CatalogDatabase catalogDatabase = new CatalogDatabaseImpl(props, dbComment);
+        return new CreateDatabaseOperation(
+                catalogManager.getCurrentCatalog(), dbName, catalogDatabase, ifNotExists);
     }
 
-    private Serializable analyzeDropDatabase(HiveParserASTNode ast) {
+    private Operation convertDropDatabase(HiveParserASTNode ast) {
         String dbName =
                 HiveParserBaseSemanticAnalyzer.unescapeIdentifier(ast.getChild(0).getText());
         boolean ifExists = false;
@@ -841,25 +1141,36 @@ public class HiveParserDDLSemanticAnalyzer {
             ifCascade = true;
         }
 
-        return new HiveParserDropDatabaseDesc(dbName, ifExists, ifCascade);
+        return new DropDatabaseOperation(
+                catalogManager.getCurrentCatalog(), dbName, ifExists, ifCascade);
     }
 
-    private Serializable analyzeSwitchDatabase(HiveParserASTNode ast) {
+    private Operation convertSwitchDatabase(HiveParserASTNode ast) {
         String dbName =
                 HiveParserBaseSemanticAnalyzer.unescapeIdentifier(ast.getChild(0).getText());
-        SwitchDatabaseDesc switchDatabaseDesc = new SwitchDatabaseDesc(dbName);
-        return new DDLWork(new HashSet<>(), new HashSet<>(), switchDatabaseDesc);
+        return new UseDatabaseOperation(catalogManager.getCurrentCatalog(), dbName);
     }
 
-    private Serializable analyzeDropTable(HiveParserASTNode ast, TableType expectedType) {
+    private Operation convertDropTable(HiveParserASTNode ast, TableType expectedType) {
         String tableName =
                 HiveParserBaseSemanticAnalyzer.getUnescapedName(
                         (HiveParserASTNode) ast.getChild(0));
         boolean ifExists = (ast.getFirstChildWithType(HiveASTParser.TOK_IFEXISTS) != null);
 
-        boolean ifPurge = (ast.getFirstChildWithType(HiveASTParser.KW_PURGE) != null);
-        return new HiveParserDropTableDesc(
-                tableName, expectedType == TableType.VIRTUAL_VIEW, ifExists, ifPurge);
+        ObjectIdentifier identifier = parseObjectIdentifier(tableName);
+        CatalogBaseTable baseTable = getCatalogBaseTable(identifier, true);
+
+        if (expectedType == TableType.VIRTUAL_VIEW) {
+            if (baseTable instanceof CatalogTable) {
+                throw new ValidationException("DROP VIEW for a table is not allowed");
+            }
+            return new DropViewOperation(identifier, ifExists, false);
+        } else {
+            if (baseTable instanceof CatalogView) {
+                throw new ValidationException("DROP TABLE for a view is not allowed");
+            }
+            return new DropTableOperation(identifier, ifExists, false);
+        }
     }
 
     private void validateAlterTableType(
@@ -893,7 +1204,7 @@ public class HiveParserDDLSemanticAnalyzer {
         }
     }
 
-    private Serializable analyzeAlterTableProps(
+    private Operation convertAlterTableProps(
             String[] qualified,
             HashMap<String, String> partSpec,
             HiveParserASTNode ast,
@@ -939,26 +1250,61 @@ public class HiveParserDDLSemanticAnalyzer {
                 }
             }
         }
-        HiveParserAlterTableDesc alterTblDesc = null;
         if (isUnset) {
             handleUnsupportedOperation("Unset properties not supported");
-        } else {
-            alterTblDesc =
-                    HiveParserAlterTableDesc.alterTableProps(
-                            tableName, partSpec, mapProp, expectView);
         }
 
-        return alterTblDesc;
+        if (expectView) {
+            return convertAlterViewProps(tableName, mapProp);
+        } else {
+            Map<String, String> newProps = new HashMap<>();
+            newProps.put(ALTER_TABLE_OP, CHANGE_TBL_PROPS.name());
+            newProps.putAll(mapProp);
+            return convertAlterTableProps(tableName, partSpec, newProps);
+        }
     }
 
-    private Serializable analyzeAlterTableSerdeProps(
+    private Operation convertAlterTableProps(
+            String tableName, Map<String, String> partSpec, Map<String, String> newProps) {
+        ObjectIdentifier tableIdentifier = parseObjectIdentifier(tableName);
+        CatalogBaseTable catalogBaseTable = getCatalogBaseTable(tableIdentifier);
+        if (catalogBaseTable instanceof CatalogView) {
+            throw new ValidationException("ALTER TABLE for a view is not allowed");
+        }
+        CatalogTable oldTable = (CatalogTable) catalogBaseTable;
+        CatalogPartitionSpec catalogPartitionSpec =
+                partSpec != null ? new CatalogPartitionSpec(partSpec) : null;
+        CatalogPartition catalogPartition =
+                partSpec != null ? getPartition(tableIdentifier, catalogPartitionSpec) : null;
+
+        Map<String, String> props = new HashMap<>();
+        if (catalogPartition != null) {
+            props.putAll(catalogPartition.getProperties());
+            props.putAll(newProps);
+            return new AlterPartitionPropertiesOperation(
+                    tableIdentifier,
+                    catalogPartitionSpec,
+                    new CatalogPartitionImpl(props, catalogPartition.getComment()));
+        } else {
+            props.putAll(oldTable.getOptions());
+            props.putAll(newProps);
+            return new AlterTableOptionsOperation(tableIdentifier, oldTable.copy(props));
+        }
+    }
+
+    private Operation convertAlterTableSerdeProps(
             HiveParserASTNode ast, String tableName, HashMap<String, String> partSpec) {
         HashMap<String, String> mapProp =
                 getProps((HiveParserASTNode) (ast.getChild(0)).getChild(0));
-        return HiveParserAlterTableDesc.alterSerDe(tableName, partSpec, null, mapProp);
+        Map<String, String> newProps = new HashMap<>();
+        newProps.put(ALTER_TABLE_OP, CHANGE_SERDE_PROPS.name());
+        for (String key : mapProp.keySet()) {
+            newProps.put(SERDE_INFO_PROP_PREFIX + key, mapProp.get(key));
+        }
+        return convertAlterTableProps(tableName, partSpec, newProps);
     }
 
-    private Serializable analyzeAlterTableSerde(
+    private Operation convertAlterTableSerde(
             HiveParserASTNode ast, String tableName, HashMap<String, String> partSpec) {
         String serdeName =
                 HiveParserBaseSemanticAnalyzer.unescapeSQLString(ast.getChild(0).getText());
@@ -966,10 +1312,18 @@ public class HiveParserDDLSemanticAnalyzer {
         if (ast.getChildCount() > 1) {
             mapProp = getProps((HiveParserASTNode) (ast.getChild(1)).getChild(0));
         }
-        return HiveParserAlterTableDesc.alterSerDe(tableName, partSpec, serdeName, mapProp);
+        Map<String, String> newProps = new HashMap<>();
+        newProps.put(ALTER_TABLE_OP, CHANGE_SERDE_PROPS.name());
+        newProps.put(SERDE_LIB_CLASS_NAME, serdeName);
+        if (mapProp != null) {
+            for (String key : mapProp.keySet()) {
+                newProps.put(SERDE_INFO_PROP_PREFIX + key, mapProp.get(key));
+            }
+        }
+        return convertAlterTableProps(tableName, partSpec, newProps);
     }
 
-    private Serializable analyzeAlterTableFileFormat(
+    private Operation convertAlterTableFileFormat(
             HiveParserASTNode ast, String tableName, HashMap<String, String> partSpec)
             throws SemanticException {
 
@@ -980,18 +1334,21 @@ public class HiveParserDDLSemanticAnalyzer {
             throw new ValidationException("Unknown AST node for ALTER TABLE FILEFORMAT: " + child);
         }
 
-        HiveParserAlterTableDesc alterTblDesc =
-                HiveParserAlterTableDesc.alterFileFormat(tableName, partSpec);
-        alterTblDesc.setGenericFileFormatName(format.getGenericName());
-
-        return alterTblDesc;
+        Map<String, String> newProps = new HashMap<>();
+        newProps.put(ALTER_TABLE_OP, CHANGE_FILE_FORMAT.name());
+        newProps.put(STORED_AS_FILE_FORMAT, format.getGenericName());
+        return convertAlterTableProps(tableName, partSpec, newProps);
     }
 
-    private Serializable analyzeAlterTableLocation(
+    private Operation convertAlterTableLocation(
             HiveParserASTNode ast, String tableName, HashMap<String, String> partSpec) {
         String newLocation =
                 HiveParserBaseSemanticAnalyzer.unescapeSQLString(ast.getChild(0).getText());
-        return HiveParserAlterTableDesc.alterLocation(tableName, partSpec, newLocation);
+        Map<String, String> newProps = new HashMap<>();
+        newProps.put(ALTER_TABLE_OP, CHANGE_LOCATION.name());
+        newProps.put(TABLE_LOCATION_URI, newLocation);
+
+        return convertAlterTableProps(tableName, partSpec, newProps);
     }
 
     public static HashMap<String, String> getProps(HiveParserASTNode prop) {
@@ -1111,13 +1468,26 @@ public class HiveParserDDLSemanticAnalyzer {
         }
     }
 
+    private CatalogPartition getPartition(
+            ObjectIdentifier tableIdentifier, CatalogPartitionSpec partitionSpec) {
+        return catalogManager
+                .getPartition(tableIdentifier, partitionSpec)
+                .orElseThrow(
+                        () ->
+                                new ValidationException(
+                                        String.format(
+                                                "Partition %s of table %s doesn't exist",
+                                                partitionSpec.getPartitionSpec(),
+                                                tableIdentifier)));
+    }
+
     /**
      * A query like this will generate a tree as follows "describe formatted default.maptable
      * partition (b=100) id;" TOK_TABTYPE TOK_TABNAME --> root for tablename, 2 child nodes mean DB
      * specified default maptable TOK_PARTSPEC --> root node for partition spec. else columnName
      * TOK_PARTVAL b 100 id --> root node for columnName formatted
      */
-    private Serializable analyzeDescribeTable(HiveParserASTNode ast) throws SemanticException {
+    private Operation convertDescribeTable(HiveParserASTNode ast) throws SemanticException {
         HiveParserASTNode tableTypeExpr = (HiveParserASTNode) ast.getChild(0);
 
         String dbName = null;
@@ -1157,37 +1527,19 @@ public class HiveParserDDLSemanticAnalyzer {
             handleUnsupportedOperation("DESCRIBE COLUMNS is not supported");
         }
 
-        DescTableDesc descTblDesc = new DescTableDesc(getResFile(), tableName, partSpec, colPath);
-
+        boolean isExt = false;
+        boolean isFormatted = false;
         if (ast.getChildCount() == 2) {
             int descOptions = ast.getChild(1).getType();
-            descTblDesc.setFormatted(descOptions == HiveASTParser.KW_FORMATTED);
-            descTblDesc.setExt(descOptions == HiveASTParser.KW_EXTENDED);
+            isExt = descOptions == HiveASTParser.KW_EXTENDED;
+            isFormatted = descOptions == HiveASTParser.KW_FORMATTED;
             if (descOptions == HiveASTParser.KW_PRETTY) {
                 handleUnsupportedOperation("DESCRIBE PRETTY is not supported.");
             }
         }
 
-        return new DDLWork(getInputs(), getOutputs(), descTblDesc);
-    }
-
-    private Serializable analyzeDescDatabase(HiveParserASTNode ast) {
-
-        boolean isExtended;
-        String dbName;
-
-        if (ast.getChildCount() == 1) {
-            dbName = HiveParserBaseSemanticAnalyzer.stripQuotes(ast.getChild(0).getText());
-            isExtended = false;
-        } else if (ast.getChildCount() == 2) {
-            dbName = HiveParserBaseSemanticAnalyzer.stripQuotes(ast.getChild(0).getText());
-            isExtended = true;
-        } else {
-            throw new ValidationException("Unexpected Tokens at DESCRIBE DATABASE");
-        }
-
-        DescDatabaseDesc descDbDesc = new DescDatabaseDesc(getResFile(), dbName, isExtended);
-        return new DDLWork(getInputs(), getOutputs(), descDbDesc);
+        ObjectIdentifier tableIdentifier = parseObjectIdentifier(tableName);
+        return new DescribeTableOperation(tableIdentifier, isExt || isFormatted);
     }
 
     public static HashMap<String, String> getPartSpec(HiveParserASTNode partspec) {
@@ -1209,8 +1561,22 @@ public class HiveParserDDLSemanticAnalyzer {
         return partSpec;
     }
 
-    private Serializable analyzeShowPartitions(HiveParserASTNode ast) throws SemanticException {
-        ShowPartitionsDesc showPartsDesc;
+    // Get the partition specs from the tree
+    private List<Map<String, String>> getPartitionSpecs(CommonTree ast) {
+        List<Map<String, String>> partSpecs = new ArrayList<>();
+        // get partition metadata if partition specified
+        for (int childIndex = 0; childIndex < ast.getChildCount(); childIndex++) {
+            HiveParserASTNode partSpecNode = (HiveParserASTNode) ast.getChild(childIndex);
+            // sanity check
+            if (partSpecNode.getType() == HiveASTParser.TOK_PARTSPEC) {
+                Map<String, String> partSpec = getPartSpec(partSpecNode);
+                partSpecs.add(partSpec);
+            }
+        }
+        return partSpecs;
+    }
+
+    private Operation convertShowPartitions(HiveParserASTNode ast) throws SemanticException {
         String tableName =
                 HiveParserBaseSemanticAnalyzer.getUnescapedName(
                         (HiveParserASTNode) ast.getChild(0));
@@ -1224,23 +1590,19 @@ public class HiveParserDDLSemanticAnalyzer {
 
         validateTable(tableName, null);
 
-        showPartsDesc = new ShowPartitionsDesc(tableName, getResFile(), partSpec);
-        return new DDLWork(getInputs(), getOutputs(), showPartsDesc);
-    }
-
-    private Serializable analyzeShowDatabases(HiveParserASTNode ast) {
-        ShowDatabasesDesc showDatabasesDesc;
-        if (ast.getChildCount() == 1) {
-            String databasePattern =
-                    HiveParserBaseSemanticAnalyzer.unescapeSQLString(ast.getChild(0).getText());
-            showDatabasesDesc = new ShowDatabasesDesc(getResFile(), databasePattern);
-        } else {
-            showDatabasesDesc = new ShowDatabasesDesc(getResFile());
+        ObjectIdentifier tableIdentifier = parseObjectIdentifier(tableName);
+        CatalogPartitionSpec spec = null;
+        if (partSpec != null && !partSpec.isEmpty()) {
+            spec = new CatalogPartitionSpec(new HashMap<>(partSpec));
         }
-        return new DDLWork(getInputs(), getOutputs(), showDatabasesDesc);
+        return new ShowPartitionsOperation(tableIdentifier, spec);
     }
 
-    private Serializable analyzeShowTables(HiveParserASTNode ast, boolean expectView) {
+    private Operation convertShowDatabases() {
+        return new ShowDatabasesOperation();
+    }
+
+    private Operation convertShowTables(HiveParserASTNode ast, boolean expectView) {
         String dbName = currentDB;
         String pattern = null;
 
@@ -1276,7 +1638,7 @@ public class HiveParserDDLSemanticAnalyzer {
         if (pattern != null) {
             handleUnsupportedOperation("SHOW TABLES/VIEWS LIKE is not supported");
         }
-        return new HiveParserShowTablesDesc(null, dbName, expectView);
+        return expectView ? new ShowViewsOperation() : new ShowTablesOperation();
     }
 
     /**
@@ -1285,46 +1647,15 @@ public class HiveParserDDLSemanticAnalyzer {
      *
      * @param ast The parsed command tree.
      */
-    private Serializable analyzeShowFunctions(HiveParserASTNode ast) {
-        ShowFunctionsDesc showFuncsDesc;
-        if (ast.getChildCount() == 1) {
-            String funcNames =
-                    HiveParserBaseSemanticAnalyzer.stripQuotes(ast.getChild(0).getText());
-            showFuncsDesc = new ShowFunctionsDesc(getResFile(), funcNames);
-        } else if (ast.getChildCount() == 2) {
+    private Operation convertShowFunctions(HiveParserASTNode ast) {
+        if (ast.getChildCount() == 2) {
             assert (ast.getChild(0).getType() == HiveASTParser.KW_LIKE);
             throw new ValidationException("SHOW FUNCTIONS LIKE is not supported yet");
-        } else {
-            showFuncsDesc = new ShowFunctionsDesc(getResFile());
         }
-        return new DDLWork(getInputs(), getOutputs(), showFuncsDesc);
+        return new ShowFunctionsOperation();
     }
 
-    /**
-     * Add the task according to the parsed command tree. This is used for the CLI command "DESCRIBE
-     * FUNCTION;".
-     *
-     * @param ast The parsed command tree.
-     */
-    private Serializable analyzeDescFunction(HiveParserASTNode ast) {
-        String funcName;
-        boolean isExtended;
-
-        if (ast.getChildCount() == 1) {
-            funcName = HiveParserBaseSemanticAnalyzer.stripQuotes(ast.getChild(0).getText());
-            isExtended = false;
-        } else if (ast.getChildCount() == 2) {
-            funcName = HiveParserBaseSemanticAnalyzer.stripQuotes(ast.getChild(0).getText());
-            isExtended = true;
-        } else {
-            throw new ValidationException("Unexpected Tokens at DESCRIBE FUNCTION");
-        }
-
-        DescFunctionDesc descFuncDesc = new DescFunctionDesc(getResFile(), funcName, isExtended);
-        return new DDLWork(getInputs(), getOutputs(), descFuncDesc);
-    }
-
-    private Serializable analyzeAlterTableRename(
+    private Operation convertAlterTableRename(
             String[] source, HiveParserASTNode ast, boolean expectView) throws SemanticException {
         String[] target =
                 HiveParserBaseSemanticAnalyzer.getQualifiedTableName(
@@ -1332,12 +1663,16 @@ public class HiveParserDDLSemanticAnalyzer {
 
         String sourceName = HiveParserBaseSemanticAnalyzer.getDotName(source);
         String targetName = HiveParserBaseSemanticAnalyzer.getDotName(target);
+        ObjectIdentifier objectIdentifier = parseObjectIdentifier(sourceName);
+        checkAlterTableLegal(objectIdentifier, expectView);
 
-        return HiveParserAlterTableDesc.rename(sourceName, targetName, expectView);
+        return expectView
+                ? new AlterViewRenameOperation(objectIdentifier, parseObjectIdentifier(targetName))
+                : new AlterTableRenameOperation(
+                        objectIdentifier, parseObjectIdentifier(targetName));
     }
 
-    private Serializable analyzeAlterTableRenameCol(
-            String[] qualified, HiveParserASTNode ast, HashMap<String, String> partSpec)
+    private Operation convertAlterTableRenameCol(String[] qualified, HiveParserASTNode ast)
             throws SemanticException {
         String newComment = null;
         boolean first = false;
@@ -1388,23 +1723,41 @@ public class HiveParserDDLSemanticAnalyzer {
         }
 
         String tblName = HiveParserBaseSemanticAnalyzer.getDotName(qualified);
-        return HiveParserAlterTableDesc.changeColumn(
-                tblName,
-                HiveParserBaseSemanticAnalyzer.unescapeIdentifier(oldColName),
-                HiveParserBaseSemanticAnalyzer.unescapeIdentifier(newColName),
-                newType,
-                newComment,
-                first,
-                flagCol,
-                isCascade);
+
+        ObjectIdentifier tableIdentifier = parseObjectIdentifier(tblName);
+        CatalogBaseTable catalogBaseTable = getCatalogBaseTable(tableIdentifier);
+        if (catalogBaseTable instanceof CatalogView) {
+            throw new ValidationException("ALTER TABLE for a view is not allowed");
+        }
+        CatalogTable oldTable = (CatalogTable) catalogBaseTable;
+        String oldName = HiveParserBaseSemanticAnalyzer.unescapeIdentifier(oldColName);
+        String newName = HiveParserBaseSemanticAnalyzer.unescapeIdentifier(newColName);
+
+        if (oldTable.getPartitionKeys().contains(oldName)) {
+            // disallow changing partition columns
+            throw new ValidationException("CHANGE COLUMN cannot be applied to partition columns");
+        }
+        TableSchema oldSchema = oldTable.getSchema();
+        TableColumn newTableColumn =
+                TableColumn.physical(
+                        newName,
+                        HiveTypeUtil.toFlinkType(TypeInfoUtils.getTypeInfoFromTypeString(newType)));
+        TableSchema newSchema =
+                OperationConverterUtils.changeColumn(
+                        oldSchema, oldName, newTableColumn, first, flagCol);
+        Map<String, String> props = new HashMap<>(oldTable.getOptions());
+        props.put(ALTER_TABLE_OP, ALTER_COLUMNS.name());
+        if (isCascade) {
+            props.put(ALTER_COL_CASCADE, "true");
+        }
+        return new AlterTableSchemaOperation(
+                tableIdentifier,
+                new CatalogTableImpl(
+                        newSchema, oldTable.getPartitionKeys(), props, oldTable.getComment()));
     }
 
-    private Serializable analyzeAlterTableModifyCols(
-            String[] qualified,
-            HiveParserASTNode ast,
-            HashMap<String, String> partSpec,
-            boolean replace)
-            throws SemanticException {
+    private Operation convertAlterTableModifyCols(
+            String[] qualified, HiveParserASTNode ast, boolean replace) throws SemanticException {
 
         String tblName = HiveParserBaseSemanticAnalyzer.getDotName(qualified);
         List<FieldSchema> newCols =
@@ -1414,10 +1767,69 @@ public class HiveParserDDLSemanticAnalyzer {
             isCascade = true;
         }
 
-        return HiveParserAlterTableDesc.addReplaceColumns(tblName, newCols, replace, isCascade);
+        ObjectIdentifier tableIdentifier = parseObjectIdentifier(tblName);
+        CatalogBaseTable catalogBaseTable = getCatalogBaseTable(tableIdentifier);
+        if (catalogBaseTable instanceof CatalogView) {
+            throw new ValidationException("ALTER TABLE for a view is not allowed");
+        }
+        CatalogTable oldTable = (CatalogTable) catalogBaseTable;
+
+        // prepare properties
+        Map<String, String> props = new HashMap<>(oldTable.getOptions());
+        props.put(ALTER_TABLE_OP, ALTER_COLUMNS.name());
+        if (isCascade) {
+            props.put(ALTER_COL_CASCADE, "true");
+        }
+        TableSchema oldSchema = oldTable.getSchema();
+        final int numPartCol = oldTable.getPartitionKeys().size();
+        TableSchema.Builder builder = TableSchema.builder();
+        // add existing non-part col if we're not replacing
+        if (!replace) {
+            List<TableColumn> nonPartCols =
+                    oldSchema.getTableColumns().subList(0, oldSchema.getFieldCount() - numPartCol);
+            for (TableColumn column : nonPartCols) {
+                builder.add(column);
+            }
+            setWatermarkAndPK(builder, oldSchema);
+        }
+        // add new cols
+        for (FieldSchema col : newCols) {
+            builder.add(
+                    TableColumn.physical(
+                            col.getName(),
+                            HiveTypeUtil.toFlinkType(
+                                    TypeInfoUtils.getTypeInfoFromTypeString(col.getType()))));
+        }
+        // add part cols
+        List<TableColumn> partCols =
+                oldSchema
+                        .getTableColumns()
+                        .subList(oldSchema.getFieldCount() - numPartCol, oldSchema.getFieldCount());
+        for (TableColumn column : partCols) {
+            builder.add(column);
+        }
+        return new AlterTableSchemaOperation(
+                tableIdentifier,
+                new CatalogTableImpl(
+                        builder.build(),
+                        oldTable.getPartitionKeys(),
+                        props,
+                        oldTable.getComment()));
     }
 
-    private Serializable analyzeAlterTableDropParts(
+    private static void setWatermarkAndPK(TableSchema.Builder builder, TableSchema schema) {
+        for (WatermarkSpec watermarkSpec : schema.getWatermarkSpecs()) {
+            builder.watermark(watermarkSpec);
+        }
+        schema.getPrimaryKey()
+                .ifPresent(
+                        pk -> {
+                            builder.primaryKey(
+                                    pk.getName(), pk.getColumns().toArray(new String[0]));
+                        });
+    }
+
+    private Operation convertAlterTableDropParts(
             String[] qualified, HiveParserASTNode ast, boolean expectView) {
 
         boolean ifExists = ast.getFirstChildWithType(HiveASTParser.TOK_IFEXISTS) != null;
@@ -1441,16 +1853,24 @@ public class HiveParserDDLSemanticAnalyzer {
 
         validateAlterTableType(tab, AlterTableDesc.AlterTableTypes.DROPPARTITION, expectView);
 
-        return new DropPartitionDesc(qualified[0], qualified[1], partSpecs, ifExists);
+        ObjectIdentifier tableIdentifier =
+                catalogManager.qualifyIdentifier(
+                        UnresolvedIdentifier.of(qualified[0], qualified[1]));
+        CatalogBaseTable catalogBaseTable = getCatalogBaseTable(tableIdentifier);
+        if (catalogBaseTable instanceof CatalogView) {
+            throw new ValidationException("DROP PARTITION for a view is not supported");
+        }
+        List<CatalogPartitionSpec> specs =
+                partSpecs.stream().map(CatalogPartitionSpec::new).collect(Collectors.toList());
+        return new DropPartitionsOperation(tableIdentifier, ifExists, specs);
     }
 
     /**
      * Add one or more partitions to a table. Useful when the data has been copied to the right
      * location by some other process.
      */
-    private Serializable analyzeAlterTableAddParts(
-            String[] qualified, CommonTree ast, boolean expectView) throws SemanticException {
-
+    private Operation convertAlterTableAddParts(
+            String[] qualified, CommonTree ast, boolean expectView) {
         // ^(TOK_ALTERTABLE_ADDPARTS identifier ifNotExists?
         // alterStatementSuffixAddPartitionsElement+)
         boolean ifNotExists = ast.getChild(0).getType() == HiveASTParser.TOK_IFNOTEXISTS;
@@ -1466,14 +1886,19 @@ public class HiveParserDDLSemanticAnalyzer {
         Map<String, String> currentPart = null;
         // Parser has done some verification, so the order of tokens doesn't need to be verified
         // here.
-        AddPartitionDesc addPartitionDesc =
-                new AddPartitionDesc(tab.getDbName(), tab.getTableName(), ifNotExists);
+        List<CatalogPartitionSpec> specs = new ArrayList<>();
+        List<CatalogPartition> partitions = new ArrayList<>();
         for (int num = start; num < numCh; num++) {
             HiveParserASTNode child = (HiveParserASTNode) ast.getChild(num);
             switch (child.getToken().getType()) {
                 case HiveASTParser.TOK_PARTSPEC:
                     if (currentPart != null) {
-                        addPartitionDesc.addPartition(currentPart, currentLocation);
+                        specs.add(new CatalogPartitionSpec(currentPart));
+                        Map<String, String> props = new HashMap<>();
+                        if (currentLocation != null) {
+                            props.put(TABLE_LOCATION_URI, currentLocation);
+                        }
+                        partitions.add(new CatalogPartitionImpl(props, null));
                         currentLocation = null;
                     }
                     currentPart = getPartSpec(child);
@@ -1495,25 +1920,96 @@ public class HiveParserDDLSemanticAnalyzer {
 
         // add the last one
         if (currentPart != null) {
-            addPartitionDesc.addPartition(currentPart, currentLocation);
+            specs.add(new CatalogPartitionSpec(currentPart));
+            Map<String, String> props = new HashMap<>();
+            if (currentLocation != null) {
+                props.put(TABLE_LOCATION_URI, currentLocation);
+            }
+            partitions.add(new CatalogPartitionImpl(props, null));
         }
 
-        return new DDLWork(getInputs(), getOutputs(), addPartitionDesc);
+        ObjectIdentifier tableIdentifier =
+                tab.getDbName() == null
+                        ? parseObjectIdentifier(tab.getTableName())
+                        : catalogManager.qualifyIdentifier(
+                                UnresolvedIdentifier.of(tab.getDbName(), tab.getTableName()));
+        CatalogBaseTable catalogBaseTable = getCatalogBaseTable(tableIdentifier);
+        if (catalogBaseTable instanceof CatalogView) {
+            throw new ValidationException("ADD PARTITION for a view is not supported");
+        }
+        return new AddPartitionsOperation(tableIdentifier, ifNotExists, specs, partitions);
     }
 
-    // Get the partition specs from the tree
-    private List<Map<String, String>> getPartitionSpecs(CommonTree ast) {
-        List<Map<String, String>> partSpecs = new ArrayList<>();
-        // get partition metadata if partition specified
-        for (int childIndex = 0; childIndex < ast.getChildCount(); childIndex++) {
-            HiveParserASTNode partSpecNode = (HiveParserASTNode) ast.getChild(childIndex);
-            // sanity check
-            if (partSpecNode.getType() == HiveASTParser.TOK_PARTSPEC) {
-                Map<String, String> partSpec = getPartSpec(partSpecNode);
-                partSpecs.add(partSpec);
+    private Operation convertAlterViewProps(String tableName, Map<String, String> newProps) {
+        ObjectIdentifier viewIdentifier = parseObjectIdentifier(tableName);
+        CatalogBaseTable baseTable = getCatalogBaseTable(viewIdentifier);
+        if (baseTable instanceof CatalogTable) {
+            throw new ValidationException("ALTER VIEW for a table is not allowed");
+        }
+        CatalogView oldView = (CatalogView) baseTable;
+        Map<String, String> props = new HashMap<>(oldView.getOptions());
+        props.putAll(newProps);
+        CatalogView newView =
+                new CatalogViewImpl(
+                        oldView.getOriginalQuery(),
+                        oldView.getExpandedQuery(),
+                        oldView.getSchema(),
+                        props,
+                        oldView.getComment());
+        return new AlterViewPropertiesOperation(viewIdentifier, newView);
+    }
+
+    private void checkAlterTableLegal(ObjectIdentifier objectIdentifier, boolean expectView) {
+        CatalogBaseTable catalogBaseTable = getCatalogBaseTable(objectIdentifier);
+        if (expectView) {
+            if (catalogBaseTable instanceof CatalogTable) {
+                throw new ValidationException("ALTER VIEW for a table is not allowed");
+            }
+        } else {
+            if (catalogBaseTable instanceof CatalogView) {
+                throw new ValidationException("ALTER TABLE for a view is not allowed");
             }
         }
-        return partSpecs;
+    }
+
+    private ObjectIdentifier parseObjectIdentifier(String compoundName) {
+        UnresolvedIdentifier unresolvedIdentifier = hiveParser.parseIdentifier(compoundName);
+        return catalogManager.qualifyIdentifier(unresolvedIdentifier);
+    }
+
+    private CatalogDatabase getDatabase(String databaseName) {
+        Catalog catalog = catalogManager.getCatalog(catalogManager.getCurrentCatalog()).get();
+        CatalogDatabase database;
+        try {
+            database = catalog.getDatabase(databaseName);
+        } catch (DatabaseNotExistException e) {
+            throw new ValidationException(String.format("Database %s not exists", databaseName), e);
+        }
+        return database;
+    }
+
+    private CatalogBaseTable getCatalogBaseTable(ObjectIdentifier tableIdentifier) {
+        return getCatalogBaseTable(tableIdentifier, false);
+    }
+
+    private CatalogBaseTable getCatalogBaseTable(
+            ObjectIdentifier tableIdentifier, boolean ifExists) {
+        Optional<CatalogManager.TableLookupResult> optionalCatalogTable =
+                catalogManager.getTable(tableIdentifier);
+        if (!optionalCatalogTable.isPresent()) {
+            if (ifExists) {
+                return null;
+            } else {
+                throw new ValidationException(
+                        String.format(
+                                "Table or View %s doesn't exist.", tableIdentifier.toString()));
+            }
+        }
+        if (optionalCatalogTable.get().isTemporary()) {
+            throw new ValidationException(
+                    String.format("Table or View %s is temporary.", tableIdentifier.toString()));
+        }
+        return optionalCatalogTable.get().getTable();
     }
 
     /**
@@ -1560,10 +2056,6 @@ public class HiveParserDDLSemanticAnalyzer {
             }
         }
         return retValue;
-    }
-
-    private Path getResFile() {
-        return SessionState.getLocalSessionPath(conf);
     }
 
     private static void handleUnsupportedOperation(HiveParserASTNode astNode) {
